extract CircularProgress component into its own file and update HomeScreen to use it
