+++ components/CircularProgressBar.tsx
@@ -1,0 +1,73 @@
+import { View, Text, Animated, Dimensions } from 'react-native';
+import Svg, { Circle } from 'react-native-svg';
+import { useEffect, useRef } from 'react';
+
+const { width } = Dimensions.get('window');
+
+const AnimatedCircle = Animated.createAnimatedComponent(Circle);
+
+interface CircularProgressProps {
+  progress: number;
+  totalDosage: number;
+  completedDosage: number;
+}
+
+export function CircularProgress({
+  progress,
+  totalDosage,
+  completedDosage,
+}: CircularProgressProps) {
+  const animatedValue = useRef(new Animated.Value(0)).current;
+  const size = width * 0.55;
+  const strokeWidth = 15;
+  const radius = (size - strokeWidth) / 2;
+  const circumference = 2 * Math.PI * radius;
+
+  useEffect(() => {
+    Animated.timing(animatedValue, {
+      toValue: progress,
+      duration: 1500,
+      useNativeDriver: true,
+    }).start();
+  }, [progress]);
+
+  const strokeDashoffset = animatedValue.interpolate({
+    inputRange: [0, 1],
+    outputRange: [circumference, 0],
+  });
+
+  return (
+    <View className="flex items-center justify-center my-2">
+      <View className="absolute flex items-center justify-center z-10">
+        <Text className="text-4xl font-bold text-white">
+          {Math.round(progress)}%
+        </Text>
+        <Text className="text-sm text-white/80 mt-1">
+          {completedDosage} of {totalDosage} completed
+        </Text>
+      </View>
+      <Svg width={size} height={size}>
+        <Circle
+          cx={size / 2}
+          cy={size / 2}
+          r={radius}
+          stroke="rgba(255, 255, 255, 0.2)"
+          strokeWidth={strokeWidth}
+          fill="none"
+        />
+        <AnimatedCircle
+          cx={size / 2}
+          cy={size / 2}
+          r={radius}
+          stroke="white"
+          strokeWidth={strokeWidth}
+          fill="none"
+          strokeDasharray={circumference}
+          strokeDashoffset={strokeDashoffset}
+          strokeLinecap="round"
+          transform={`rotate(-90 ${size / 2} ${size / 2})`}
+        />
+      </Svg>
+    </View>
+  );
+}
